requirements.txt
asgiref==3.7.2
boto3==1.28.31
botocore==1.31.31
dj-database-url==2.1.0
Django==4.2.4
django-crispy-forms==2.0
django-storages==1.13.2
environs==9.5.0
gunicorn==21.2.0
jmespath==1.0.1
marshmallow==3.20.1
packaging==23.1
Pillow==10.0.0
psycopg2==2.9.7
python-dateutil==2.8.2
python-dotenv==1.0.0
s3transfer==0.6.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
whitenoise==6.5.0
